requirements.txt
librosa
pandas
pydub
numpy
matplotlib
tensorflow
pillow
scikit-learn
